--- Code/LLDSA/Entities/FileOperator/LlsFileProcessorFactory.cs
@@ -1,16 +1,0 @@
-﻿using System.Dynamic;
-using LLDSA.Entities.FileOperator.LLSFileProcessor;
-
-namespace LLDSA.Entities.FileOperator
-{
-    public class LlsFileProcessorFactory
-    {
-
-        public static LlsFileProcessor GetLlsFileProcessor()
-        {
-            //Todo 
-            return null;
-        }
-
-    }
-}
